refactor(test): fix failed tests
